# docs/public/api/openapi-3.0.yaml
openapi: 3.0.0
info:
  description: Track, Manage, and Organize your Things.
  title: Homebox API
  contact:
    name: Homebox Team
    url: https://discord.homebox.software
  version: "1.0"
paths:
...
  /v1/entities:
    get:
      security:
        - Bearer: []
      tags:
        - Entities
      summary: Query All Entities
      parameters:
        - description: search string
          name: q
          in: query
          schema:
            type: string
        - description: page number
          name: page
          in: query
          schema:
            type: integer
        - description: items per page
          name: pageSize
          in: query
          schema:
            type: integer
        - description: tags Ids
          name: tags
          in: query
          explode: true
          schema:
            type: array
            items:
              type: string
        - description: parent Ids
          name: parentIds
          in: query
          explode: true
          schema:
            type: array
            items:
              type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/repo.EntityListResult"
    post:
      security:
        - Bearer: []
      tags:
        - Entities
      summary: Create Entity
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/repo.EntityCreate"
        description: Entity Data
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/repo.EntityOut"
  /v1/entities/bulk-edit:
    get:
      security:
        - Bearer: []
      tags:
        - Entities
      summary: Query entities with all fields required by the bulk editor
      parameters:
        - description: search string
          name: q
          in: query
          schema:
            type: string
        - description: page number
          name: page
          in: query
          schema:
            type: integer
        - description: items per page
          name: pageSize
          in: query
          schema:
            type: integer
        - description: tags Ids
          name: tags
          in: query
          explode: true
          schema:
            type: array
            items:
              type: string
        - description: parent Ids
          name: parentIds
          in: query
          explode: true
          schema:
            type: array
            items:
              type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/repo.EntityBulkEditListResult"
  /v1/entities/export:
    get:
      security:
        - Bearer: []
      tags:
        - Entities
      summary: Export Entities
      responses:
        "200":
          description: text/csv
          content:
            "*/*":
              schema:
                type: string
  /v1/entities/fields:
    get:
      security:
        - Bearer: []
      tags:
        - Entities
      summary: Get All Custom Field Names
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /v1/entities/fields/values:
    get:
      security:
        - Bearer: []
      tags:
        - Entities
      summary: Get All Custom Field Values
      parameters:
        - description: Field name
          name: field
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /v1/entities/import:
    post:
      security:
        - Bearer: []
      tags:
        - Entities
      summary: Import Entities
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                csv:
                  description: CSV file to upload
                  type: string
                  format: binary
              required:
                - csv
        required: true
      responses:
        "204":
          description: No Content
  /v1/entities/tree:
    get:
      security:
        - Bearer: []
      tags:
        - Entities
      summary: Get Locations Tree
      parameters:
        - description: include items in response tree
          name: withItems
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/repo.TreeItem"
  "/v1/entities/{id}":
    get:
      security:
        - Bearer: []
      tags:
        - Entities
      summary: Get Entity
      parameters:
        - description: Entity ID
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/repo.EntityOut"
    put:
      security:
        - Bearer: []
      tags:
        - Entities
      summary: Update Entity
      parameters:
        - description: Entity ID
          name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/repo.EntityUpdate"
        description: Entity Data
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/repo.EntityOut"
    delete:
      security:
        - Bearer: []
      tags:
        - Entities
      summary: Delete Entity
      parameters:
        - description: Entity ID
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
    patch:
      security:
        - Bearer: []
      tags:
        - Entities
      summary: Patch Entity
      parameters:
        - description: Entity ID
          name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/repo.EntityPatch"
        description: Entity Data
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/repo.EntityOut"
...
  "/v1/entities/{id}/duplicate":
    post:
      security:
        - Bearer: []
      tags:
        - Entities
      summary: Duplicate Entity
      parameters:
        - description: Entity ID
          name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/repo.DuplicateOptions"
        description: Duplicate Options
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/repo.EntityOut"
...
  "/v1/entities/{id}/path":
    get:
      security:
        - Bearer: []
      tags:
        - Entities
      summary: Get the full path of an entity
      parameters:
        - description: Entity ID
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/repo.EntityPath"
...
servers:
  - url: /api
components:
  securitySchemes:
    Bearer:
      description: "\"Type 'Bearer TOKEN' to correctly set the API Key\""
      type: apiKey
      name: Authorization
      in: header
  schemas:
...
    repo.DuplicateOptions:
      type: object
      properties:
        copyAttachments:
          type: boolean
        copyCustomFields:
          type: boolean
        copyMaintenance:
          type: boolean
        copyPrefix:
          type: string
    repo.EntityBulkEditListResult:
      type: object
      properties:
        fieldDefinitions:
          type: array
          items:
            $ref: "#/components/schemas/repo.EntityFieldDefinition"
        items:
          type: array
          items:
            $ref: "#/components/schemas/repo.EntityOut"
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
        totalPrice:
          type: number
    repo.EntityCreate:
      type: object
      required:
        - name
      properties:
        description:
          type: string
          maxLength: 1000
        entityTypeId:
          type: string
        manufacturer:
          type: string
          maxLength: 255
          x-omitempty: true
          nullable: true
        modelNumber:
          description: |-
            Identifications — optional at create time; populated e.g. by the
            barcode product-search import flow (#1578).
          type: string
          maxLength: 255
          x-omitempty: true
          nullable: true
        name:
          type: string
          maxLength: 255
          minLength: 1
        parentId:
          type: string
          nullable: true
        quantity:
          type: number
        tagIds:
          description: Edges
          type: array
          items:
            type: string
...
    repo.EntityListResult:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/repo.EntitySummary"
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
        totalPrice:
          type: number
    repo.EntityOut:
      type: object
      properties:
        archived:
          type: boolean
        assetId:
          type: string
          example: "0"
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/repo.ItemAttachment"
        children:
          description: Container-specific fields (for entities whose
            entity_type.is_location = true)
          type: array
          items:
            $ref: "#/components/schemas/repo.EntitySummary"
        createdAt:
          type: string
        description:
          type: string
        entityType:
          allOf:
            - $ref: "#/components/schemas/repo.EntityTypeSummary"
          x-omitempty: true
          nullable: true
        fields:
          type: array
          items:
            $ref: "#/components/schemas/repo.EntityFieldData"
        id:
          type: string
        imageId:
          type: string
          x-omitempty: true
          nullable: true
        insured:
          type: boolean
        itemCount:
          description: Container-specific (populated when querying locations)
          type: number
        lifetimeWarranty:
          description: Warranty
          type: boolean
        location:
          description: |-
            Location is the nearest ancestor whose entity type is a location.
            When the direct parent is already a location it equals Parent; when
            the entity is nested inside other items it is the location those
            items ultimately live in. Nil for top-level entities.
          allOf:
            - $ref: "#/components/schemas/repo.EntitySummary"
          x-omitempty: true
          nullable: true
        manufacturer:
          type: string
        modelNumber:
          type: string
        name:
          type: string
        notes:
          description: Extras
          type: string
        parent:
          description: Edges
          allOf:
            - $ref: "#/components/schemas/repo.EntitySummary"
          x-omitempty: true
          nullable: true
        purchaseDate:
          description: Purchase
          type: string
        purchaseFrom:
          type: string
        purchasePrice:
          type: number
        quantity:
          type: number
        serialNumber:
          type: string
        soldDate:
          description: Sold
          type: string
        soldNotes:
          type: string
        soldPrice:
          type: number
        soldTo:
          type: string
        syncChildEntityLocations:
          type: boolean
        tags:
          type: array
          items:
            $ref: "#/components/schemas/repo.TagSummary"
        thumbnailId:
          type: string
          x-omitempty: true
          nullable: true
        totalPrice:
          type: number
        updatedAt:
          type: string
        warrantyDetails:
          type: string
        warrantyExpires:
          type: string
...